pyproject.toml
[tool.poetry]
name = "genai-wordcup-042024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "genai_wordcup_042024"}]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.0.37"
langchain = "^0.1.16"
environ-config = "^23.2.0"
langchain-openai = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
